+++ src/features/lost/components/main/ItemCard.styles.ts
