+++ src/features/lost/components/main/Registration.styles.ts
