+++ src/features/lost/components/main/StaffLabel.types.ts
