+++ src/features/lost/index.ts
@@ -1,6 +1,6 @@
-export { default as ItemList } from './components/ItemList';
-export { default as Registration } from './components/Registration';
-export { default as ItemCard } from './components/ItemCard';
+export { default as ItemList } from './components/main/ItemList';
+export { default as Registration } from './components/main/Registration';
+export { default as ItemCard } from './components/main/ItemCard';
 export { default as ModalNotification } from './modal/ModalNotification';
 export { default as ModalPost } from './modal/ModalPost';
-export { default as StaffLabel } from './components/StaffLabel';
+export { default as StaffLabel } from './components/main/StaffLabel';
